node: track single npm global versions, add pnpm and bun

+++ .zsh/plugins/npm-global/npm-global.plugin.zsh
@@ -3,44 +3,36 @@
 function npm-global() {
   # Default global directory
   [[ -z "$NPM_GLOBAL" ]] && export NPM_GLOBAL="$HOME/.npm"
-  [[ ! -d "$NPM_GLOBAL" ]] && mkdir -p "$NPM_GLOBAL"
+  mkdir -p "$NPM_GLOBAL"
 
-  ##
-  # Prints the current node version.
-  #
-  local function node_version() {
-    cd ~
-    echo ${$(node -v)#v}
-  }
+  # Define commands that trigger a dump
+  local DUMP_COMMANDS=(
+    install add i in ins inst insta instal isnt isntal isntall
+    uninstall unlink remove rm r
+    un update up upgrade udpate
+  )
 
   ##
   # Prints a list of the globally install dependencies.
   #
-  local function global_dependencies() {
+  local function global_deps() {
     command npm list -g --depth=0 | sed -e '1d' -e 's/^[^ ]* //' | tr '\n' ' ' | sed 's/[[:space:]]*$//'
   }
 
   ##
-  # Prints a list of the locally install dependencies.
-  #
-  local function local_dependencies() {
-    jq -r '.dependencies | keys[]' package.json 2>/dev/null
-  }
-
-  ##
-  # Dumps global dependencies to the version-specific package.json located in NPM_GLOBAL.
+  # Dumps global dependencies to the version-specific package.json located in $NPM_GLOBAL.
   #
   function npm-dump() {
-    local dir=$NPM_GLOBAL/$(node_version)
-    mkdir -p $dir
-    cd $dir
-
-    if [[ -f package.json ]]; then
-      command npm uninstall $(jq -r '.dependencies | keys[]' package.json | tr '\n' ' ')
+    if [[ ! -z $1 ]]; then
+      echo "Unknown arguments: $@"
+      echo
+      echo "To dump global dependencies, run:"
+      echo "  npm dump"
+      return 1
     fi
-
-    echo "{\n  \"engines\": {\n    \"node\": \"$(node_version)\"\n  }\n}" > package.json
-    command npm install $(global_dependencies)
+    cd $NPM_GLOBAL
+    echo "{}" > package.json
+    command npm install $(global_deps) --package-lock-only
     local exit=$?
     rm -rf node_modules
     cd - >/dev/null
@@ -48,18 +40,11 @@ function npm-global() {
   }
 
   ##
-  # Background process to dump global dependencies.
-  #
-  local function npm-background-dump() {
-    ((npm-dump) &>/dev/null &) &>/dev/null
-  }
-
-  ##
   # Updates and syncs global npm packages.
   #
   function npm-fresh() {
-    if [[ ! -z "${@:2}" ]]; then
-      echo "Unknown arguments: ${@:2}"
+    if [[ ! -z $1 ]]; then
+      echo "Unknown arguments: $@"
       echo
       echo "To update all global dependencies, run:"
       echo "  npm fresh"
@@ -67,7 +52,7 @@ function npm-global() {
     fi
     command npm -g update
     local exit=$?
-    npm-background-dump
+    ((npm-dump) &>/dev/null &) &>/dev/null
     return $exit
   }
 
@@ -87,11 +72,8 @@ function npm-global() {
 
     if [[ $global == false ]]; then
       case $1 in
-        dump)
-          npm-dump
-          ;;
-        fresh)
-          npm-fresh
+        dump|fresh)
+          npm-$1 ${@:2}
           ;;
         *)
           command npm $@
@@ -103,14 +85,9 @@ function npm-global() {
     command npm $@
     local exit=$?
 
-    local DUMP_COMMANDS=(
-      install add i in ins inst insta instal isnt isntal isntall
-      uninstall unlink remove rm r
-      un update up upgrade udpate
-    )
     for cmd in $DUMP_COMMANDS; do
       if [[ $@ == *$cmd* ]]; then
-        npm-background-dump
+        ((npm-dump) &>/dev/null &) &>/dev/null
         break
       fi
     done
@@ -124,7 +101,7 @@ function npm-global() {
 
     case $1 in
       install)
-        ((npm install $(global_dependencies) --global && npm-dump) &>/dev/null &) &>/dev/null
+        ((npm install $(global_deps) --global && npm-dump) &>/dev/null &) &>/dev/null
         ;;
     esac
